+++ docs/_/@stores/GithubStore/index.js
