+++ docs/src/entities/Article-stability--browser-spec.jsx
@@ -1,13 +1,20 @@
 import React from 'react';
 import { mount } from 'enzyme';
+import { Provider } from 'mobx-react';
+import GithubStore from '@stores/GithubStore';
 import Article from './Article';
 
 describe('Article', () => {
   Article.all.forEach((article) => {
     describe(article.path(), () => {
       it('renders', () => {
+        const githubStore = new GithubStore();
         const Component = article.component();
-        mount(<Component />);
+        mount(
+          <Provider github={githubStore}>
+          <Component />
+          </Provider>
+      );
       });
     });
   });
