+++ package.json
@@ -76,6 +76,7 @@
     "react-dom": "^15.6.1",
     "react-router": "^3.0.5",
     "split-array": "^1.0.1",
+    "whatwg-fetch": "^2.0.3",
     "xojs": "^0.5.0"
   },
   "devDependencies": {
